+++ app/src/main/java/com/example/master/models/User.kt
@@ -4,7 +4,7 @@ data class User(
     val email: String = "",
     val fullName: String = "",
     val age: Int = 0,
-    val weight: Float = 0f,
+    val weight: Int = 0,
     val height: Int = 0,
     val shareData: Boolean = true
 )
